% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesofman-package.R
\docType{package}
\name{agesofman-package}
\alias{agesofman}
\alias{agesofman-package}
\title{agesofman: Remote Sensing Statistical Applications in Agriculture}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Ages of Man project applies advanced statistical methodology on remote sensing agriculture.
}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} [thesis advisor]
}

}
\keyword{internal}
