% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\alias{create,PersephoneFiles-method}
\title{Create new objects}
\usage{
create(files, ...)

\S4method{create}{PersephoneFiles}(files)
}
\arguments{
\item{files}{an S4 object of class \code{Files}.}

\item{...}{extra arguments.}

\item{dir}{character. The parent directory in which the files are stored.}
}
\value{
An S4 object of class "Project"ModelList
}
\description{
Function \code{create()} takes a \code{files} argument that
contains the all the required file names to build a new model. The files need
to be inside a directory following the tree structure of \code{hermes}.
}
\details{
The files have to be .R scripts that create an object named as per
the file type name. That is, the prm file should create a list called
\code{prm}, the dts file a list called \code{dts} and so on. The scripts
are evaluated and the objects are used in the model construction. The
\code{prm} list should also contain an elements called \code{Class},
specifying the model class. Note that the combined lists should cover all
the necessary slots of the class (the ones without a prototype).
}
\examples{
\dontrun{
files <- PersephoneFiles(prm = "test", dts = "test")
object <- create(files)
}
}
