% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes.R
\name{download_hermes}
\alias{download_hermes}
\title{Download project \code{hermes}}
\usage{
download_hermes(dir = getwd())
}
\arguments{
\item{dir}{character. The parent directory inside which the project will
be downloaded}
}
\value{
nothing. The project is downloaded in the specified directory.
}
\description{
Download project \code{hermes} from the github repository.
}
\examples{
\dontrun{
download_hermes()
}
}
