% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{load_agesofman}
\alias{load_agesofman}
\title{Automatically load agesofman}
\usage{
load_agesofman(level = "project")
}
\arguments{
\item{level}{character. The type of profile, one of 'user' or 'project'.}
}
\value{
nothing. The .Renviron file is edited.
}
\description{
Automatically load package agesofman by editing the
.Renviron file.
}
\examples{
\dontrun{
load_agesofman()
}
}
