% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus-package.R
\docType{package}
\name{prometheus-package}
\alias{prometheus}
\alias{prometheus-package}
\title{prometheus: Remote Sensing Statistical Applications in Agriculture}
\description{
Prometheus is a family of R packages that apply statistical methodologies on agricultural problems utilizing remote sensing data.
}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} [thesis advisor]
}

}
\keyword{internal}
