% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_files.R
\docType{class}
\name{PrometheusFiles-class}
\alias{PrometheusFiles-class}
\alias{PrometheusFiles}
\alias{PersephoneFiles-class}
\alias{PersephoneFiles}
\title{Prometheus File Classes}
\value{
An S4 object of class \code{PrometheusFiles} or one of its
subclasses.
}
\description{
Prometheus File Classes
}
\section{Slots}{

\describe{
\item{\code{prm}}{character. The parameter file name.}

\item{\code{dts}}{character. The dataset file name.}
}}

\examples{
\dontrun{
files <- PersephoneFiles(prm = "test", dts = "test")
}
}
