% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_exe}
\alias{read_exe}
\alias{read_prm}
\alias{read_dts}
\title{Read scripts}
\usage{
read_exe(name, type, project, dir = getwd())

read_prm(name, project, dir = getwd())

read_dts(name, project, dir = getwd())
}
\arguments{
\item{name}{character. The name of the script (without the .R suffix).}

\item{type}{character. The type of the script}

\item{project}{character. The project the script concerns.}

\item{dir}{character. The hermes project directory.}
}
\value{
read_exe sources the script, while read_prm and read_dts source
the script locally and only return the prm and dts objects, respectively.
}
\description{
Read scripts
}
\examples{
\dontrun{
read_prm("test", "persephone")
}
}
