% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_startup.R
\name{set_agesofman}
\alias{set_agesofman}
\title{Set agesofman}
\usage{
set_agesofman(path_demeter, path_hermes)
}
\arguments{
\item{path_demeter}{character. The path to the database directory.}

\item{path_hermes}{character. The path to \code{hermes} project root directory.}
}
\value{
nothing. The .Renviron file is edited.
}
\description{
Load package agesofman at startup and add default path_demeter
and path_hermes by editing the .Renviron file.
}
\examples{
\dontrun{
# Set path
path_hermes <- getwd()

# Download project
download_hermes(path_hermes)

# Set default path
set_path_hermes(path = path_hermes)
get_path_hermes()
set_agesofman(path_demeter = getwd(), path_hermes = path_hermes)
}
}
