% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{set_path_hermes}
\alias{set_path_hermes}
\alias{get_path_hermes}
\title{Set hermes path}
\usage{
set_path_hermes(path = NULL, level = "project")

get_path_hermes()
}
\arguments{
\item{path}{character. The path to the database directory.}

\item{level}{character. The level of profile, one of 'user' or 'project'.}
}
\value{
nothing. The .Renviron file is edited.
}
\description{
Set or get the default \code{path_hermes} by editing the .Renviron
file.
}
\examples{
\dontrun{
set_path_demeter(getwd(), level = "user")
get_path_demeter()
}
}
\seealso{
\code{\link[cronus:get_startup_path]{cronus::get_startup_path()}}
}
