% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_exe}
\alias{read_exe}
\alias{read_prm}
\alias{read_dts}
\title{Read scripts}
\usage{
read_exe(name, type, project, dir = getwd())

read_prm(name, project, dir = getwd())

read_dts(name, project, dir = getwd())
}
\arguments{
\item{name}{character. The name of the script (without the .R suffix).}

\item{type}{character. The type of the script}

\item{project}{character. The project the script concerns.}

\item{dir}{character. The hermes project directory.}
}
\description{
This set of functions sources scripts inside the structured
directory of the \code{hermes} project and returns an appropriate object. These
functions are used in the model creation process.
}
\section{Functions}{
\itemize{
\item \code{read_exe()}: source the script.

\item \code{read_prm()}: source the script locally and return object \code{prm}.

\item \code{read_dts()}: source the script locally and return object \code{dts}.

}}
\examples{
\dontrun{
read_prm("test", "persephone")
}
}
