% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{set_agesofman}
\alias{set_agesofman}
\title{Set agesofman}
\usage{
set_agesofman(path_demeter, path_hermes = NULL, level = "project")
}
\arguments{
\item{path_demeter}{character. The path to the database directory.}

\item{path_hermes}{character. The path to \code{hermes} project root directory.}

\item{level}{character. The type of profile, one of 'user' or 'project'.}
}
\value{
nothing. The .Renviron file is edited.
}
\description{
Load package agesofman at startup and add default path_demeter
and path_hermes by editing the .Renviron file.
}
\examples{
\dontrun{
set_agesofman(path_demeter = getwd())
}
}
